move meta string to meta dir
