+++ cpp/fory/util/buffer.h
@@ -95,8 +95,9 @@ public:
     reinterpret_cast<T *>(data_ + offset)[0] = value;
   }
 
-  template <typename T,
-            typename = meta::EnableIfIsOneOf<T, int8_t, uint8_t, bool>>
+  template <typename T, typename = std::enable_if_t<std::disjunction_v<
+                            std::is_same<T, int8_t>, std::is_same<T, uint8_t>,
+                            std::is_same<T, bool>>>>
   inline void UnsafePutByte(uint32_t offset, T value) {
     data_[offset] = value;
   }
@@ -113,8 +114,9 @@ public:
     return value;
   }
 
-  template <typename T,
-            typename = meta::EnableIfIsOneOf<T, int8_t, uint8_t, bool>>
+  template <typename T, typename = std::enable_if_t<std::disjunction_v<
+                            std::is_same<T, int8_t>, std::is_same<T, uint8_t>,
+                            std::is_same<T, bool>>>>
   inline T GetByteAs(uint32_t relative_offset) {
     FORY_CHECK(relative_offset < size_) << "Out of range " << relative_offset
                                         << " should be less than " << size_;
